requirements.txt
streamlit>=1.29.0
openai>=1.6.1
PyPDF2>=3.0.0
google-auth==2.37.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.156.0
requests>=2.32.3
